Update git prompt on chdir
When switching directory using the Location mode, the
$edit:after-command handlers aren't run, so I need to hook into the
$before-chdir to make this work.
It's not the most performant way of doing things, but it works, and
I've yet to notice any real lag.

I could do some introspection in the $edit:after-command handler, to
not update the gitstatus when the command run was a `cd`, but frankly,
that seems like an uphill battle.