Add iter module with reduce function
Iteration functions are handy, so having a dedicated module for them is
probably a good idea. This reduce implementation uses the value
channel, meaning it streams the output, which is pretty dope.
The one weakness is, that it doesn't have any way of acting upon
arbitrary input, ala $input? from the stdlib. Not sure how to do that.